chore: improve Select exports
